Completed the leet conversion; Refactoring

+++ wordListGenerator.py
@@ -7,13 +7,6 @@ from lib import write_file as wf
 from lib import usage as us
 from lib import combination as co
 
-# myWord = Word("Pierre")
-# myWord.addLeet()
-# myWord.addUpperCase()
-# myWord.addLowerCase()
-# myWord.loadWordList()
-# myWord.printWordList()
-
 def main(argv):
 
     writeFilePath = ''
@@ -21,9 +14,10 @@ def main(argv):
     leet = False
     lowercase = False
     uppercase = False
+    simulation = False
 
     try:                                
-        opts, args = getopt.getopt(argv, "w:k:lCc", ["write=", "keywords=", "leet", "uppercase", "lowercase"])
+        opts, args = getopt.getopt(argv, "w:k:lCcs", ["write=", "keywords=", "leet", "uppercase", "lowercase", "simulation"])
     except getopt.GetoptError:          
         us.usage()                         
         sys.exit(2)
@@ -42,6 +36,8 @@ def main(argv):
             lowercase = True
         elif opt in ("-C", "--uppercase"):
             uppercase = True
+        elif opt in ("-s", "--simulation"):
+            simulation = True
             
     if writeFilePath == '':
         print("Error: -w/--write option required")
@@ -56,9 +52,10 @@ def main(argv):
     keyWordFile.read()
     keyWordFile.loadKeyWord()
     keyWordList = keyWordFile.returnKeyWord()
-    print(keyWordList)
     globalTab = []
     
+    weight = 1
+    tmpWeight = 0
     for possibilities in keyWordList:
         wordsTab = []
         for string in possibilities:
@@ -70,58 +67,23 @@ def main(argv):
             if (uppercase == True):
                 myWord.addUpperCase()
             myWord.loadNumbers()
+            tmpWeight = tmpWeight + myWord.weightPossibilities()
             wordsTab.append(myWord)
+        weight = weight * tmpWeight
         globalTab.append(wordsTab)
             
     myCombination = co.Combination(globalTab)
     myCombination.loadNumbers()
     nb = myCombination.returnNbCombination()
     
-    file = wf.WriteFile(writeFilePath)
-    for i in range(nb):
-        file.write(myCombination.convertNumberInCombination(i)+"\n")
-    file.close()
-   
-   
-    # 
-    # for line in result:
-        # file.write(line+"\n")
-    # file.close()
-    
-def test():
-    myWord = wd.Word("1111")
-    myWord.addLeet()
-    myWord.loadNumbers()
-    nb = myWord.returnNbCombination()
-    myWord2 = wd.Word("2222")
-    myWord2.addLeet()
-    myWord2.loadNumbers()
-    nb2 = myWord2.returnNbCombination()
-    myWord3 = wd.Word("1111")
-    myWord3.addLeet()
-    myWord3.loadNumbers()
-    nb3 = myWord3.returnNbCombination()
-    myWord4 = wd.Word("2222")
-    myWord4.addLeet()
-    myWord4.loadNumbers()
-    nb4 = myWord4.returnNbCombination()
-    myCombination = co.Combination([[myWord, myWord2],[myWord3, myWord4]])
-    myCombination.loadNumbers()
-    nb5 = myCombination.returnNbCombination()
-    
-    print(nb)
-    print(nb2)
-    print(nb3)
-    print(nb4)
-    print(nb5)
-    
-    
-    print(myCombination.convertNumberInCombination(1562499))
-    # myWord.addUpperCase()
-    # myWord.addLowerCase()
-    # myWord.loadWordList()
-    # myWord.printWordList()
+    if simulation == False:
+        file = wf.WriteFile(writeFilePath)
+        for i in range(nb):
+            file.write(myCombination.convertNumberInCombination(i)+"\n")
+        file.close()
+    if simulation == True:
+        print(str(nb) + " Combinations")
+        # print(str(weight / 1024) + " Ko")
 
 if __name__ == "__main__":
-    # test()
     main(sys.argv[1:])
